Finished refactor for Flet 0.84.0
