--- src/default_config.py
@@ -1,177 +1,0 @@
-DEFAULT_CONFIG = """# === ProductivityApp Config ===
-# Lines starting with "#" are comments.
-# Add more items to each list as you like.
-# Restart app to apply changes.
-
-[PRODUCTIVE]
-# Common productivity tools and IDEs
-apps = [
-    "code.exe",
-    "pycharm64.exe",
-    "idea64.exe",
-    "clion64.exe",
-    "rider64.exe",
-    "devenv.exe",
-    "sublime_text.exe",
-    "notepad++.exe",
-    "obsidian.exe",
-    "anki.exe",
-    "notion.exe",
-    "onenote.exe",
-    "excel.exe",
-    "winword.exe",
-    "mspowerpoint.exe",
-    "powerbi.exe",
-    "acrobat.exe",
-    "pdfreader.exe",
-    "wps.exe",
-    "todoist.exe",
-    "ticktick.exe",
-    "clockify.exe",
-    "slack.exe",
-    "teams.exe",
-    "zoom.exe",
-    "outlook.exe",
-    "thunderbird.exe",
-    "postman.exe",
-    "datagrip.exe",
-    "dbeaver.exe",
-    "mysqlworkbench.exe",
-    "studio64.exe",        # IntelliJ variants
-    "xmind.exe",
-    "drawio.exe",
-    "notepad.exe"
-]
-
-# Extra check for productive window titles / keywords
-keywords = [
-    # Study & learning
-    "dashboard",
-    "course",
-    "lecture",
-    "tutorial",
-    "study",
-    "exam",
-    "notes",
-    "assignment",
-    "research",
-    "reference",
-    "textbook",
-    "math",
-    "mathematics",
-    "calculus",
-    "physics",
-    "science",
-    "biology",
-    "chemistry",
-    "project report",
-    "thesis",
-    "google docs",
-    "google sheets",
-    "visual studio code",
-    "pycharm",
-    "intellij",
-    "jetbrains",
-    "data analysis",
-    "data visualization",
-    "spreadsheet",
-    "report",
-    "kanban board",
-    "jira",
-    "asana",
-    "trello",
-    "git",
-    "github",
-    "gitlab",
-    "documentation",
-    "stack overflow",
-    "stackexchange",
-    "openai",
-    "api reference",
-    "python script",
-    "programming",
-    "development",
-    "debug",
-    "run configuration",
-    "terminal",
-    "command prompt",
-    "powershell",
-    "vscode"
-]
-
-[DISTRACTING]
-# Apps that are likely distracting or non-productive
-apps = [
-    "tlauncher.exe",
-    "roblox.exe",
-    "steam.exe",
-    "epicgameslauncher.exe",
-    "genshinimpact.exe",
-    "valorant.exe",
-    "leagueclient.exe",
-    "leagueoflegends.exe",
-    "minecraftlauncher.exe",
-    "discord.exe",
-    "spotify.exe",
-    "netflix.exe",
-    "twitch.exe",
-    "chrome.exe",          # fuzzy-matched against social media titles
-    "msedge.exe",
-    "opera.exe",
-    "opera_gx.exe",
-    "firefox.exe"
-]
-
-# Social media, streaming, or other distracting content
-keywords = [
-    "youtube",
-    "facebook",
-    "instagram",
-    "reddit",
-    "twitter",
-    "tiktok",
-    "discord",
-    "messenger",
-    "netflix",
-    "twitch",
-    "spotify",
-    "soundcloud",
-    "pinterest",
-    "9gag",
-    "tumblr",
-    "snapchat",
-    "roblox",
-    "steam",
-    "epic games",
-    "genshin",
-    "valorant",
-    "league of legends",
-    "minecraft",
-    "whatsapp",
-    "telegram",
-    "hoyolab",
-    "threads",
-    "kick",
-    "prime video",
-    "hbo",
-    "disney plus",
-    "crunchyroll",
-    "bilibili",
-    "dailymotion",
-    "deviantart",
-    "anime",
-    "manga",
-    "movie",
-    "series",
-    "livestream",
-    "chat",
-    "music",
-    "playlist",
-    "funny videos",
-    "shorts",
-    "gaming",
-    "let's play",
-    "stream",
-    "watching now"
-]
-"""
