+++ src/setup.py
@@ -1,11 +1,8 @@
 import flet as ft
-import asyncio
 from typing import Optional
-
 
 WINDOW_WIDTH: Optional[ft.Number] = 550
 WINDOW_HEIGHT: Optional[ft.Number] = 400
-
 
 def before_main_ui(page: ft.Page):
     """Use for the `before_main` in `run()`."""
@@ -28,20 +25,3 @@ def before_main_ui(page: ft.Page):
     page.window.on_event = lambda e: print(e)
     page.update()
     
-async def fix_stretched_window(
-    page: ft.Page, *,
-    center_page: bool = False
-):
-    """
-    When launching a Flet desktop app, sometimes the window appears to be stretched.
-    The fix? Just resize it. So, that's exactly what this does.
-    """
-    page.window.width = WINDOW_WIDTH * 1.1
-    page.window.height = WINDOW_HEIGHT * 1.1
-    page.window.update()
-    await asyncio.sleep(1)
-    page.window.width = WINDOW_WIDTH
-    page.window.height = WINDOW_HEIGHT
-    page.window.update()
-    if center_page:
-        await page.window.center()
